% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cont_rate.R
\name{cont.rate}
\alias{cont.rate}
\title{Continous Rate}
\usage{
cont.rate(r, t)
}
\arguments{
\item{r}{rate (nominal)}

\item{t}{number of compounding period}
}
\value{
Generates Continuously Compounded Rate
}
\description{
Converts nominal rate into Continously compounded Rate
}
\details{
Generates Continuously Compounded Rate
}
\examples{
cont.rate(0.025, 4)
}
