% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/put_delta.R
\name{put.delta}
\alias{put.delta}
\title{Put Delta}
\usage{
put.delta(s, k, t, sd, r, d = 0)
}
\arguments{
\item{s}{Spot Price of Underlying Asset}

\item{k}{Exercise Price of Contract}

\item{t}{Time to Expiration}

\item{sd}{Volatality}

\item{r}{Risk free rate of return}

\item{d}{Divident Yield (use cont.rate()), Default: 0}
}
\value{
Oupput gives the delta of a Option Contract.
}
\description{
Calculate the Delta (Option Greek) of a Contract
}
\details{
Delta represents the rate of change between the option's price and a $1 change in the underlying asset's price.
}
\examples{
put.delta(100, 105, 0.25, 0.35, 0.0488)
}
