% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/put_gamma.R
\name{put.gamma}
\alias{put.gamma}
\title{Put Gamma}
\usage{
put.gamma(s, k, t, sd, r, d = 0)
}
\arguments{
\item{s}{Spot Price of Underlying Asset}

\item{k}{Exercise Price of Contract}

\item{t}{Time to Expiration}

\item{sd}{Volatality}

\item{r}{Risk free rate of return}

\item{d}{Divident Yield (use cont.rate()), Default: 0}
}
\value{
Output gives the Gamma of a Option Contract.
}
\description{
Calculate the Gamma (Option Greek) of a Contract
}
\details{
Gamma represents the rate of change between an option's delta and the underlying asset's price.
}
\examples{
put.gamma(100, 105, 0.25, 0.35, 0.0488)
}
