% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/put_rho.R
\name{put.rho}
\alias{put.rho}
\title{Put Rho}
\usage{
put.rho(s, k, t, sd, r, d = 0)
}
\arguments{
\item{s}{Spot Price of Underlying Asset}

\item{k}{Exercise Price of Contract}

\item{t}{Time to Expiration}

\item{sd}{Volatality}

\item{r}{Risk free rate of return}

\item{d}{Divident Yield (use cont.rate()), Default: 0}
}
\value{
Output gives the Estimated Premium of a Option Contract.
}
\description{
Calculate the Rho (Option Greek) of Option Contract
}
\details{
Rho represents the rate of change between an option's value and a 1\% change in the interest rate.
}
\examples{
put.rho(100, 105, 0.25, 0.35, 0.0488)
}
