% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/put_spread.R
\name{put.spread}
\alias{put.spread}
\title{Bull/Bear Put Spread Strategy Function}
\usage{
put.spread(k1, k2, long_put, short_put, llimit = 20, ulimit = 20)
}
\arguments{
\item{k1}{Excercise Price of Long Put Option}

\item{k2}{Excercise Price of Short Put Option}

\item{long_put}{Premium of Long Put Option}

\item{short_put}{Premium of Short Put Option}

\item{llimit}{Lower limit of stock price at Expiration., Default: 20}

\item{ulimit}{Upper Limit of Stock Price at Expiration, Default: 20}
}
\value{
OUTPUT_DESCRIPTION Returns the profit/loss generated from the strategy along with the profit/loss of individual contract and an interactive graph for the same.
}
\description{
This function can be used to develop a Bull/Bear Put Strategy.
}
\details{
The strategy uses two put options to form a range consisting of a high strike price and a low strike price.
}
\examples{
put.spread(1.2, 3.2, 100, 105)
}
