% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/put_theta.R
\name{put.theta}
\alias{put.theta}
\title{Put Theta}
\usage{
put.theta(s, k, t, sd, r, d = 0)
}
\arguments{
\item{s}{Spot Price of Underlying Asset}

\item{k}{Exercise Price of Contract}

\item{t}{Time to Expiration}

\item{sd}{Volatality}

\item{r}{Risk free rate of return}

\item{d}{Divident Yield (use cont.rate()), Default: 0}
}
\value{
Output gives the Theta of a Option Contract.
}
\description{
Calculate the Theta (Option Greek) of Option Contract
}
\details{
Theta represents the rate of change between the option price and time, or time sensitivity - sometimes known as an option's time decay.
}
\examples{
put.theta(100, 105, 0.25, 0.35, 0.0488)
}
