% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/long_straddle.R
\name{straddle.long}
\alias{straddle.long}
\title{Long Straddle Strategy Function}
\usage{
straddle.long(c, p, k, ulimit = 10, llimit = 10)
}
\arguments{
\item{c}{Premium of Long call Option}

\item{p}{Premium of Long Put Option}

\item{k}{Excercise Price of Long call and Put Option}

\item{ulimit}{Upper Limit of Stock Price at Expiration, Default: 20}

\item{llimit}{Lower limit of stock price at Expiration., Default: 20}
}
\value{
OUTPUT_DESCRIPTION Returns the profit/loss generated from the strategy along with the profit/loss of individual contract and an interactive graph for the same.
}
\description{
This function can be used to develop a Long Straddle Strategy.
}
\details{
A straddle is a neutral options strategy that involves simultaneously buying both a put option and a call option for the underlying security with the same strike price and the same expiration date.
}
\examples{
straddle.long(1.2, 3.2, 100)
}
