% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/straddle_short.R
\name{straddle.short}
\alias{straddle.short}
\title{Short Straddle Strategy Function}
\usage{
straddle.short(c, p, k, ulimit = 10, llimit = 10)
}
\arguments{
\item{c}{Premium of Short call Option}

\item{p}{Premium of Short Put Option}

\item{k}{Excercise Price of Short call and Put Option}

\item{ulimit}{Upper Limit of Stock Price at Expiration, Default: 20}

\item{llimit}{Lower limit of stock price at Expiration., Default: 20}
}
\value{
OUTPUT_DESCRIPTION Returns the profit/loss generated from the strategy along with the profit/loss of individual contract and an interactive graph for the same.
}
\description{
This function can be used to develop a Short Straddle Strategy.
}
\details{
A straddle is a neutral options strategy that involves simultaneously selling both a put option and a call option for the underlying security with the same strike price and the same expiration date.
}
\examples{
straddle.short(1.2, 3.2, 100)
}
