% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/long_strangle.R
\name{strangle.long}
\alias{strangle.long}
\title{Long Strangle Strategy Function}
\usage{
strangle.long(c, p, k_call, k_put, ulimit = 10, llimit = 10)
}
\arguments{
\item{c}{Premium of Long call Option}

\item{p}{Premium of Long Put Option}

\item{k_call}{Excercise Price of Long call Option}

\item{k_put}{Excercise Price of Long Put Option}

\item{ulimit}{Upper Limit of Stock Price at Expiration, Default: 20}

\item{llimit}{Lower limit of stock price at Expiration., Default: 20}
}
\value{
OUTPUT_DESCRIPTION Returns the profit/loss generated from the strategy along with the profit/loss of individual contract and an interactive graph for the same.
}
\description{
This function can be used to develop a Long Strangle Strategy.
}
\details{
A strangle is an options strategy where the investor holds a position in both a call and a put option with different strike prices, but with the same expiration date and underlying asset.
}
\examples{
strangle.long(1.2, 3.2, 100, 105)
}
