% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/short_strangle.R
\name{strangle.short}
\alias{strangle.short}
\title{Short Strangle Strategy Function}
\usage{
strangle.short(c, p, k_call, k_put, ulimit = 10, llimit = 10)
}
\arguments{
\item{c}{Premium of Short call Option}

\item{p}{Premium of Short Put Option}

\item{k_call}{Excercise Price of Short call Option}

\item{k_put}{Excercise Price of Short Put Option}

\item{ulimit}{Upper Limit of Stock Price at Expiration, Default: 20}

\item{llimit}{Lower limit of stock price at Expiration., Default: 20}
}
\value{
OUTPUT_DESCRIPTION Returns the profit/loss generated from the strategy along with the profit/loss of individual contract and an interactive graph for the same.
}
\description{
This function can be used to develop a Short Strangle Strategy.
}
\details{
A strangle is an options strategy where the investor holds a position in both a call and a put option with different strike prices, but with the same expiration date and underlying asset.
}
\examples{
strangle.short(1.2, 3.2, 100, 105)
}
